DOH!  I fixed the off-by one bug for short-term history max while rotating, didn't do so for the non-rotating case.
This caused problems when lateness/jitter is growing, as it wasn't seen until
rotation happened.  This made things pretty unstable.

Now, it's solid as a rock (at least, it passes my 5 minute test that killed it
before this).

So, fix that, and remove some fudge I had added to try and work around this obvious
problem.

It now performs well, even if there's a 10% timestamp skew (incoming packets clock runs 10%
slower than local clock, requiring us to continually grow), AND 10% local packet loss.
In this case, we still have 0% lateness, and minimal delay.


git-svn-id: svn://svn.code.sf.net/p/iaxclient/code/trunk@350 f854266a-e20d-0410-a9be-8852184c940f