From: Mikael Magnusson <mikaelmagnusson@glocalnet.net> Subject: [Iaxclient-devel] [PATCH] Some bugfixes
This patch fixes some compilation error and some missing include files.

 lib/portmixer/px_win_wmme/px_win_wmme.c |    2 +-
 simpleclient/WinIAX/StdAfx.cpp          |    2 +-
 simpleclient/WinIAX/WinIAX.cpp          |   31 ++++++++++++++++---------------
 simpleclient/WinIAX/winiax.h            |    7 +++----
 simpleclient/iaxphone/iaxphone.cc       |    2 +-
 simpleclient/testcall/testcall.c        |   18 ++++++++++++------
 simpleclient/wx/wx.cc                   |    2 +-
 7 files changed, 35 insertions(+), 29 deletions(-)

====================

A few comments on this one..  I ended up commiting it anyway, but:

1) It didn't apply;  probably because of whitespace in WinIAX.cpp.  Actually, it was
because of line-endings; and patch -l doesn't make patch grok it.  So, I went and
manually changed (those two) files to use unix line endings, which makes the patches
apply.

2) This patch breaks "iax2slin", which also uses "file_set_file".  The reason I chose a
separate namespace for this, though, was because I was thinking of making a proper
separation of the "audio modules" and the main library;  calling the audio module's
function directly from a program breaks the model.  However, the "file" audio driver
itself is pretty much a quick hack..    I'd rather see the "file" driver completed to
make it more robust, and make a proper API, than entrench the hacky API..

3) I hope your changes to testcall don't break windows (ctype/unistd.h includes), but,
it's easy enough to change back.

-SteveK


git-svn-id: svn://svn.code.sf.net/p/iaxclient/code/trunk@457 f854266a-e20d-0410-a9be-8852184c940f