okay, this is where we want to be.
first, i've reconciled changes to libiax2/jitterbuf.* with copies 
that once live in up in lib.

second, target_extra is now configurable in the jitterbuffer.  
it defaults to the #define'd JB_TARGET_EXTRA, but can be set 
via the jb_setconf() call since a target_extra has been added
to the jb_conf struct.



git-svn-id: svn://svn.code.sf.net/p/iaxclient/code/trunk@649 f854266a-e20d-0410-a9be-8852184c940f