target_extra is now configurable in the jitterbuffer.
- it defaults to JB_TARGET_EXTRA.
- it can be set via libiaxclient_lib.c:iaxc_set_jb_target_extra( )

under the covers, iaxc_set_jb_target_extra() stores the passed value
in a static global that is passed to iax.c:iax_set_jb_target_extra()
when libiaxclient_lib.c:iaxc_initialize() is called.

iax.c similiarly stores the passed value as a static global and
passes it to the the jitterbuffer when iax.c:iax_session_new() 
calls jitterbuf.c:jb_setconf().

in jitterbuf.c, target_extra defaults to the #define'd JB_TARGET_EXTRA.

if iaxc_set_jb_target_extra() is not called or if -1 is passed to it, 
the jitterbuffer will use its default.



git-svn-id: svn://svn.code.sf.net/p/iaxclient/code/trunk@650 f854266a-e20d-0410-a9be-8852184c940f