Fix problem where media (audio, video, CNG) frames might be sent prior to the call being ACCEPTed.
When iaxclient initiates a call, it sends a NEW frame. The server end of the call must send back an ACCEPT frame prior to the call being in the "Linked" state. It is only after the call is in this Linked state that media frames may be transmitted by iaxclient. Note that media frames _are_ allowed to be transmitted prior to the call being ANSWERed (or completed in iaxclient parlance). See this thread for details:

  http://osdir.com/ml/telephony.pbx.asterisk.iaxclient.devel/2005-08/msg00021.html

This modifies iaxclient's decision logic for when it may send audio. Basically either the call has to have been answered (IAXC_CALL_STATE_COMPLETE) or the call has to be outgoing and the audio format set (which happens when the ACCEPT frame comes back from the server).
For video, we take the same approach and only allow video to go out after ACCEPT for outgoing calls and after ANSWER (complete) for incoming calls.

Without this fix, it is possible for iaxclient to send a NEW and (for example) CNG frames in short succession. The server may then receive the CNG frame first, before the NEW frame. The server sends the correct INVAL response to this call-less CNG frame which results in iaxclient destroying the call and sending a hangup event to the application. This is pretty bad, but then to make things wierder, the server will, of course, receive the NEW frame shortly after the CNG frame and respond to that with an ACCEPT frame.

C --> NEW -\/--> S
C --> CNG -/\--> S
C <-- INVAL  <-- S * server received CNG first
C <-- ACCEPT <-- S * server responds to NEW second!


git-svn-id: svn://svn.code.sf.net/p/iaxclient/code/trunk@1423 f854266a-e20d-0410-a9be-8852184c940f